made user and snippet controller

+++ src/main/java/com/example/ByteBank/service/UserService.java
@@ -2,12 +2,18 @@ package com.example.ByteBank.service;
 
 
 
+import com.example.ByteBank.exception.UserNotFoundException;
+import com.example.ByteBank.models.db.Snippet;
 import com.example.ByteBank.models.db.User;
 import com.example.ByteBank.repo.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
@@ -23,6 +29,10 @@ public class UserService {
         return userRepository.findById(id);
     }
 
+    public List<User> getAllUsers() {
+        return userRepository.findAll();
+    }
+
     public Optional<User> getByUsername(String username) {
         return userRepository.findByUsernameIgnoreCase(username);
     }
@@ -31,18 +41,31 @@ public class UserService {
         return userRepository.existsByUsernameIgnoreCase(username);
     }
 
+    public List<Snippet> getSnippetsByUserId(Long userId) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new RuntimeException("User not found"));
+        return new ArrayList<>(user.getSnippets());
+    }
+
+
     public User updateUser(Long id, User updatedUser) {
         return userRepository.findById(id)
                 .map(existingUser -> {
                     existingUser.setUsername(updatedUser.getUsername());
                     existingUser.setEmail(updatedUser.getEmail());
+                    existingUser.setProfilePictureUrl(updatedUser.getProfilePictureUrl());
                     // Update other fields as necessary
                     return userRepository.save(existingUser);
                 })
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .orElseThrow(() -> new UserNotFoundException(id));
     }
 
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+    public boolean deleteUser(Long id) {
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
